Fixed the bug in junction_virtual_atom_distance.
Used the ftug.virtual_atoms(bg) approach instead of the (bugy) average_stem_vres_atom_positions.py
This helps to avoid duplicate code.

Other occurences of average_stem_vres_atom_positions.py may need to be replaced in the future.

The new approach may be slower, if it poses a problem, we can store the results of ftug.virtual_atoms(bg) after the first call
instead of calculating it over and over again.