Recalculated average_atom_positions to include sidechains.
The atomname in the key in this file includes the type of nucleotide for sidechain atoms, but not for non-sidechain atoms.
That means that the non-sidechain atom positions are calculated as an average over all atoms, while the sidechaina tom positions are averages only over values for the corresponding nucleotide type.

The module forgi.threedee.utilities.average_stem_vres_atom_positions.py should not used any more, as with this commit its functionality is completely replaced by average_atom_positions.