Cleanup of some unused code.
Since the sampling of stats (has/ will be) moved to ernwin's StatStorage,
it is no longer necessary to carry the stat-file with forgi.

The ernwin-publication evaluates, in the supplement, different methods to fit
the helix cylinder to the pdb residues of a stem. It was shown that the
superimpose-based methods were superiour to other methods. There is no point
in keeping the not so good methods around, if we consider the maintainence cost.
Further more get_mids and get_twists now use stem_from_chains as a single method
for this fitting (which is based on Peter's code). It will be broken up in to
two PRIVATE functions for stems and coords in the future.