Changed the order of bulges.
Now multiloops are sorted by their adjacent nucleotides (define_a),
not by the lowest nt of the adjacent stems.

A change was necessary, because the sort order of mls was not consistent on python3
(during bulge-graph generation [relabel_nodes] the labelling previousely dependet
on the order of iteration through dictionaries.)
Since the numbering of ml-segments had to change anyway,
I decided to completely change it to be based on the define_a.