Refactor cutting at cofold cutpoint into seperate file.
I do not want the BulgeGraph object to be seen as mutable on
the level of the secondary structure, which is why I do not
like to have mamber functions for splitting.

In the future, it would be nice to seperate the BulgeGraph
into an object holding two classes, a Graph and a sequence
(Composition over inheritence), but I made already too many
changes in this refactoring step, so I will tackle this later.

One possible design would be

GraphWithArbitraryLabels -> GraphWithLabelsSHMIFT      immutable
    |                                   |
    V                                   V
GraphConstruction                CofoldInsertion           mutable

but such a scheme only works, if we get rid of the sequence
at this level.

So for now, I use function that modify the BulgeGraph from outside,
which is not optimal but at least the bulge_graph file is a little bit
shorter and better readable.