Create seperate MissingResidue class
Currently, this class is not used by the Sequence class internally,
but only used as an alternative way to pass missing residue information to the class
without having to memorize dictionary keys.

Further more, we can now export the missing residues as a list of such instances,
such that we can modify this list. See next commit (about transforming BulgeGraphs)
for a use of this feature.