Start module for A-Minor classification in forgi.
Based on code by Peter K. from ernwin, but highly modified.

We use the ,linear kernel by default, because it is 100 times faster than the gaussian.

TODO: consolidate code for creation of the model into forgi, to make it reproduceable.

We still need to find (using cross-validation) different hyper-parameters for different loop types.
We still need to calculatre P(I|A in seq) independently for different looptypes.